Finish building the auth plugin, setup route guards
The route guards handle checking whether or not the user the user can
access the login or register page. We prevent them from accessing the
login and register page if they are already logged in AND the page has
the 'requiresGuest' meta data.

Login.vue: on succesful login set the auth token for 2 hrs expiration
then redirect user to /newsfeed